+++ WooGym/lib/app/providers/payments/stripe_pay.dart
@@ -1,6 +1,4 @@
-//
-//  LabelCore
-//  Label StoreMax
+//  WooGym
 //
 //  Created by Anthony Gordon.
 //  2024, WooSignal Ltd. All rights reserved.
@@ -9,7 +7,6 @@
 //  Unless required by applicable law or agreed to in writing, software
 //  distributed under the License is distributed on an "AS IS" BASIS,
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//
 
 import 'package:flutter/material.dart';
 import '/bootstrap/app_helper.dart';