+++ WooGym/lib/config/payment_gateways.dart
@@ -1,3 +1,6 @@
+import 'package:flutter_app/app/providers/payments/cash_on_delivery.dart';
+import 'package:flutter_app/app/providers/payments/paypal_pay.dart';
+
 import '/app/models/payment_type.dart';
 import '/app/providers/payments/stripe_pay.dart';
 import '/bootstrap/helpers.dart';
@@ -8,12 +11,12 @@ import 'package:nylo_framework/nylo_framework.dart';
 | PAYMENT GATEWAYS
 |
 | Configure which payment gateways you want to use.
-| Docs here: https://woosignal.com/docs/app/label-storemax
+| Docs here: https://woosignal.com/docs/app/woogym
 |--------------------------------------------------------------------------
 */
 
 const appPaymentGateways = [];
-// Available: "Stripe"
+// Available: "Stripe", "PayPal", "CashOnDelivery"
 // e.g. app_payment_gateways = ["Stripe"];
 
 List<PaymentType> paymentTypeList = [
@@ -23,6 +26,22 @@ List<PaymentType> paymentTypeList = [
     description: trans("Debit or Credit Card"),
     assetImage: "dark_powered_by_stripe.png",
     pay: stripePay,
+  ),
+
+  addPayment(
+    id: 2,
+    name: "CashOnDelivery",
+    description: trans("Cash on delivery"),
+    assetImage: "cash_on_delivery.jpeg",
+    pay: cashOnDeliveryPay,
+  ),
+
+  addPayment(
+    id: 4,
+    name: "PayPal",
+    description: trans("Debit or Credit Card"),
+    assetImage: "paypal_logo.png",
+    pay: payPalPay,
   ),
 
   // e.g. add more here