+++ WooGym/lib/resources/pages/admin_class_detail_page.dart
@@ -1,3 +1,13 @@
+//  WooGym
+//
+//  Created by Anthony Gordon.
+//  2024, WooSignal Ltd. All rights reserved.
+//
+
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+
 import 'package:flutter/material.dart';
 import '/bootstrap/helpers.dart';
 import '/resources/widgets/safearea_widget.dart';