+++ WooGym/lib/resources/pages/login_page.dart
@@ -1,4 +1,4 @@
-//  Label StoreMax
+//  WooGym
 //
 //  Created by Anthony Gordon.
 //  2024, WooSignal Ltd. All rights reserved.
@@ -24,6 +24,16 @@ class LoginPage extends NyStatefulWidget<LoginController> {
 }
 
 class _LoginPageState extends NyState<LoginPage> {
+
+  String firstName = "";
+
+  // init() async {
+  //
+  //   whenEnv('production', perform: () {
+  //     firstName = "Anthony";
+  //   });
+  // }
+
   @override
   Widget build(BuildContext context) {
     ColorStyles color = ThemeColor.get(context);
@@ -39,6 +49,7 @@ class _LoginPageState extends NyState<LoginPage> {
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
+              Text("Welcome back, $firstName!"),
               StoreLogo(),
               Column(
                 mainAxisSize: MainAxisSize.max,
@@ -60,7 +71,7 @@ class _LoginPageState extends NyState<LoginPage> {
                       ),
                     ),
                     validationRules: "email",
-                    dummyData: "",
+                    dummyData: "test@aol.com",
                   ),
                   const SizedBox(
                     height: 16,
@@ -83,7 +94,8 @@ class _LoginPageState extends NyState<LoginPage> {
                       ),
                     ),
                     obscureText: true,
-                    dummyData: "",
+                    dummyData: "password1",
+                    // canSeePassword: true
                   ),
                   const SizedBox(
                     height: 16,