+++ WooGym/lib/resources/pages/no_connection_page.dart
@@ -1,4 +1,4 @@
-//  StoreMob
+//  WooGym
 //
 //  Created by Anthony Gordon.
 //  2024, WooSignal Ltd. All rights reserved.