+++ WooGym/lib/resources/widgets/toast_notification_widget.dart
@@ -1,11 +1,20 @@
+//  WooGym
+//
+//  Created by Anthony Gordon.
+//  2024, WooSignal Ltd. All rights reserved.
+//
+
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 
 class ToastNotification extends StatelessWidget {
-  const ToastNotification(ToastMeta toastMeta, {Function? onDismiss, Key? key})
+  const ToastNotification(ToastMeta toastMeta, {Function? onDismiss, super.key})
       : _toastMeta = toastMeta,
-        _dismiss = onDismiss,
-        super(key: key);
+        _dismiss = onDismiss;
 
   final Function? _dismiss;
   final ToastMeta _toastMeta;